+++ lua/cbits/hslua/hslua.h
@@ -4,8 +4,6 @@ int hslua_error(lua_State *L);
 
 int hslua_call_hs(lua_State *L);
 
-int hslua_userdata_gc(lua_State *L);
-
 void hslua_arith(lua_State *L, int op, int *status);
 
 int hslua_compare(lua_State *L, int index1, int index2, int op, int *status);